+++ src/video/SDL_RLEaccel.c
@@ -1111,20 +1111,20 @@ static int RLEAlphaSurface(SDL_Surface *surface)
         Uint8 *lastline = dst; /* end of last non-blank line */
 
         /* opaque counts are 8 or 16 bits, depending on target depth */
-#define ADD_OPAQUE_COUNTS(n, m)   \
-    if (df->BytesPerPixel == 4) { \
-        ((Uint16 *)dst)[0] = n;   \
-        ((Uint16 *)dst)[1] = m;   \
-        dst += 4;                 \
-    } else {                      \
-        dst[0] = n;               \
-        dst[1] = m;               \
-        dst += 2;                 \
+#define ADD_OPAQUE_COUNTS(n, m)           \
+    if (df->BytesPerPixel == 4) {         \
+        ((Uint16 *)dst)[0] = (Uint16)n;   \
+        ((Uint16 *)dst)[1] = (Uint16)m;   \
+        dst += 4;                         \
+    } else {                              \
+        dst[0] = (Uint8)n;                \
+        dst[1] = (Uint8)m;                \
+        dst += 2;                         \
     }
 
         /* translucent counts are always 16 bit */
 #define ADD_TRANSL_COUNTS(n, m) \
-    (((Uint16 *)dst)[0] = n, ((Uint16 *)dst)[1] = m, dst += 4)
+    (((Uint16 *)dst)[0] = (Uint16)n, ((Uint16 *)dst)[1] = (Uint16)m, dst += 4)
 
         for (y = 0; y < h; y++) {
             int runstart, skipstart;
@@ -1314,15 +1314,15 @@ static int RLEColorkeySurface(SDL_Surface *surface)
     w = surface->w;
     h = surface->h;
 
-#define ADD_COUNTS(n, m)        \
-    if (bpp == 4) {             \
-        ((Uint16 *)dst)[0] = n; \
-        ((Uint16 *)dst)[1] = m; \
-        dst += 4;               \
-    } else {                    \
-        dst[0] = n;             \
-        dst[1] = m;             \
-        dst += 2;               \
+#define ADD_COUNTS(n, m)                \
+    if (bpp == 4) {                     \
+        ((Uint16 *)dst)[0] = (Uint16)n; \
+        ((Uint16 *)dst)[1] = (Uint16)m; \
+        dst += 4;                       \
+    } else {                            \
+        dst[0] = (Uint8)n;              \
+        dst[1] = (Uint8)m;              \
+        dst += 2;                       \
     }
 
     for (y = 0; y < h; y++) {
